adds template migration
creates a template migration, 000-template, to define best practices for
developing future-proof migrations.

we've written several migrations that depend on schemas that are defined outside
of the migrations directory. if those schemas later change (such as in a later
migration), then that migration will break.

the template migration copies from Migration022 as a model for future-proof
migrations.