fixes serialization of mints with unicode (#3513)
we calculate the size of raw transactions incorrectly when those transactions
contain mints of assets with unicode characters in either the name or metadata:
we use 'sizeVarString' with the default binary encoding, but use
'writeVarString' with utf8 encoding. this results in an out of bounds write
because the binary encoding is smaller than the utf8 encoding.

fixes the out of bounds write by using the utf8 encoding when calculating size.

adds test of serialization and deserialization of a mint with unicode
characters.