feat: introduce WebSocketClient class to allow options parameter (#3527)
The issue right now is that since we have the IsomorphicWebSocket, we
lack the ability to set the options parameter for the ws WebSocket. The
primary issue is that any connections created this way will not receive
the maxPayload option, thus allowing anyone to send a message of any
size (up to the default limit set in the library which is I think
100mb). By creating this class which extends ws WebSocket and sets
default options in the constructor, we can create ws WebSockets with the
maxPayload option set.