uses unspentNoteHashes in getUnspentNotes
we use getUnspentNotes when funding transactions to find notes to spend.

the current implementation of getUnspentNotes iterates over all notes in the
wallet in order to find notes for the right asset that are confirmed and
unspent.

the unspentNoteHashes index that we added to support computing available balance
can also be used to efficiently find confirmed, unspent notes for a given asset.

implements walletDb methods for iterating over unspent notes by assetId and
sequence range.

rewrites getUnspentNotes to use the unspentNoteHashes index via the new walletDb
methods.