feat: replace segfault-handler
segfault-handler is a useful library, but requires a lot of extra
dependencies for our users since it compiles binaries on installation.
We can replace this with some napi rust functions that give us most of
what we need. This can later be extended to cover all the cases we want
it to, or split out into its own package.