adds 'available' balance to wallet (#3515)
* adds unspentNoteHashes store to walletDb (#3349)

the unspentNoteHashes datastore is an index on note hashes that supports
efficiently looking up the values of notes that are confirmed in a given
sequence range and have not been spent in another transaction.

this query pattern supports computing the 'available' balance for an account and
asset by summing the values of all notes that are confirmed and unspent.

the datastore uses a five-part compound key:
- account prefix
- asset ID
- sequence
- value
- note hash

the stored value is always null

the datastore is updated according to these rules:
- when a pending transaction is created

        - remove entries for any note that is spent in the transaction

- when a transaction is connected

        - add entries for all output notes (if they have not been spent in a
          pending transaction)
        - remove entries for any note that is spent in the transaction

- when a transaction is disconnected

        - remove entries for all output notes

- when a transaction expires

        - add entries for any notes spent in the expired transaction

this store will be used to calculate the 'available' or 'spendable' balance that
an account has for a given asset. to calculate the balance:
- determine the maximum sequence for confirmed notes
- iterate over all keys in unspentNoteHashes in the range:

        - gte: account prefix, asset id, sequence = 1
        - lt: account prefix, asset id, sequence = max sequence + 1

- sum the value stored in the fourth part of each key

* adds backfill for unspentNoteHashes (#3355)

* adds backfill for unspentNoteHashes

iterates over all notes for each account

if a note is on the chain and has not been spent the migration adds an entry to
the unspentNoteHashes datastore

* renames BigIntBEEncoding to BigU64BEEncoding

* renames migration to Migration023

* future-proofs migration for unspentNoteHashes

defines schemaNew with new datastore for unspentNoteHashes

defines schemaOld with schema and encodings for accounts and decryptedNotes

rewrites migration to use old and new stores to operate on database stores
directly instead of using Account methods

* uses createDB instead of initialized walletDb

* calculates available balance from unspent notes (#3354)

uses the unspentNoteHashes store to calculate the available balance

the available balance is the amount of an asset that an account currently has
available to spend in a new transaction. the available balance is equal to the
sum of the values of all notes for the asset that are confirmed on the chain and
have not been spent in any confirmed, unconfirmed, or pending transaction

adds 'loadUnspentNoteValues' to walletDb to iterate over the values of unspent
notes for a given asset

calculates the available balance by using 'loadUnspentNoteValues' to sum the
values of all unspent notes with confirmed sequences

* adds available balance to cli output (#3359)

* adds available balance to cli output

passes available balance through RPC endpoints

includes the 'available' balance in all balance outputs from the
'wallet:balance' and 'wallet:balances' commands

displays 'Available Balance' instead of the confirmed balance when running
'wallet:balance'

* checks available balance instead of confirmed

instead of using the confirmed balance in 'send' and the faucet service command,
uses the available balance to determine how much can be sent

* uses available balance instead of confirmed in asset selector

* encodes available balance in getBalanaces response

* bumps migration to 024

converts migration structure to follow 000-template

* regenerates fixtures for new tests