calculates deltas from unconfirmed transactions in bulk
to calculate the confirmed balance for a nasset we start with the unconfirmed
balance and subtract the balance delta for each unconfirmed transaction.

when we calculate confirmed balances for all assets we iterate over unconfirmed
transactions once for each asset.

defines 'getUnconfirmedDeltas' to sum the transaction balance deltas for all
assets in one pass over unconfirmed transactions.

defines 'getUnconfirmedDelta' to do the same for a particular asset.

adds unit test

removes outdated fixture