fixes balance calculation when confirmations >= chain length
if the length of the chain is shorter than the confirmation range we calculate a
negative number for the maximum confirmed sequence. this results in an error
when attempting to construct a key range with a negative value since the db
encoding requires a non-negative integer value.

however, the genesis block is always confirmed, so the maximum confirmed
sequence can never be less than 1.

fixes calculation of available balance by setting maxConfirmedSequence to be at
least 1.

adds unit test for this case