gets balance for correct account in promptCurrency
when getting the balance to use in the prompt in 'promptCurrency' we do not use
the name of the account passed to the function.

this results in getting the balance for the default account, which may not be
the right one.

we get the balance for the right account in the asset selector, so you can see
the issue clearly when using the asset selector for a non-default account and
using the amount prompt afterwards.