changes timestampToTransactionHash to one-to-many
it's possible for multiple transactions to have the same timestamp in a wallet.

for example, if an account receives two transactions that are on the same block,
but did not sync those transactions until they were on the chain, then both
transactions will have the block header timestamp.

this example might occur in practice if a node was offline when the transactions
were first created and/or did not receive the transactions in gossip messages
before they were added to the chain.

this might also happen for an account that is imported to a wallet after the
transactions were added to the chain. when rescanning the chain for the imported
account all transactions will use the timestamp of the block header of the block
they are found on.

we use the 'timestampToTransactionHash' index to load transactions in order for
the 'getAccountTransactions' RPC. the impact of the one-to-one datastore is that
the RPC would miss transactions that the account has.