fixes deletion of expired mint transactions
when a mint transaction expires, we delete assets that that mint created from
the wallet's assets store.

during the deletion we assert that the asset that we are about to delete is
currently in the store. when running 'wallet:prune' we fail to delete an expired
mint transaction because the asset has already been deleted from the assets
store.

it's also possible to see this error after a reorg if an account had multiple
mints on the fork: the original mint will expire first, and a subsequent mint
with the same asset will hit the error because the asset will have already been
deleted when the first mint expired.

changes behavior of 'deleteCreatedAssetsFromTransaction' to match other deletion
methods like 'deleteDecryptedNote': if the object is not found in the store,
return early from the method.