fixes scanTransactions to update ScanState after every block
scan state is updated with 'scan.signal'. we currently call 'scan.signal' within
the logic of 'connectBlock' after scanning a transaction for an account. the
result of this is that we only update the scan state for a sequence if some
account in the wallet has a transaction at that sequence. the impact of this is
that the output of `wallet:rescan' does not update as the scan progresses
through the chain.

updates the scan state within 'scanTransactions' after each call to
'connectBlock' so that the scan state reflects that the block has been connected
for all accounts.