fixes getUnspentNotes when confirmations >= chain length
if the length of the chain is shorter than or equal to the confirmation range
then we calculate a non-positive number for the maximum confirmed sequence. this
results in an error when attempting to construct a key range for loading unspent
notes from the unspentNoteHashes index.

the genesis block is always confirmed, so the maximum confirmed sequence can
never be less than 1.

fixes getUnspentNotes by setting maxConfirmedSequence to be at least 1.