fixes count of subscribed pool miners
the StratumServer has a property, 'subscribed', that it uses to count the number
of subscribed miners.

we don't decrement that counter when a subscribed miner is disconnected because
of an error. this causes the miner count in status messages to hover at an
incorrect level after miner errors.

we also use the number of clients to indicate the number of miners when sending
notifications for submitted blocks. this is not the right number since there may
be connected clients that are only following the pool for status messages and
aren't mining.