avoids deleting wallet in reset command
we likely do NOT want to delete users' accounts and keys after mainnet launch.

removes deletion of wallet database from 'reset' command. when the user starts a
node on a new network after running reset, the wallet will automatically reset
because the account heads will not be on the same chain as the network.