improves output of stop command (#3641)
the 'stopNode' RPC shuts down the node before returning a response, so the cli
client receives an error from the lost connection. IronfishCommand handles RPC
connection errors and outputs 'Cannot connect to your node, start your node
first.'. this is confusing both because we were able to connect to the node and
because we successfully stopped it.

the stop command also silences RPC connection errors, so nothing is output if
the node is already stopped when stop is run.

- returns a response from stopNode before shutting down the node
- removes error handling from 'client.connect' in stop command since the
  IronfishCommand abstract class handles these errors
- adds output to 'stop' command indicating that it asks the node to shut down