sets imported createdAt to null if not in chain
when importing an account with a non-null createdAt block we need to ensure that
that block is on the same chain as the other accounts in the wallet. if it is on
a different chain, on a fork, or ahead of the wallet's chain then we cannot
correctly scan any transactions for that account.

- on import checks whether createdAt refers to a block in the node's chain and
  sets createdAt to null in the new account if not