starts scanTransactions after account creation
if a node is online and syncing at the time of account creation it's possible to
for an account head to be set to a block that is behind the wallet's
chainProcessor.

if the account's head is behind the chain processor then the wallet will not
scan any transactions for the account until the node is restarted.

runs scanTransactions after creating an account in the 'create' RPC so that a
newly created account will be synced with the chainProcessor if it is behind.
this matches the logic in the 'importAccount' RPC.