fixes rescanning wallet without --reset (#3654)
the logic of 'connectBlock' filters any account that does not have its 'head'
equal to the previous block. this means that if we don't reset account heads
during rescan nothing will happen.

this is also true of rescanning using the '--from' flag. this flag lets users
start the rescan at a particular sequence. however, if we don't set account
heads to that sequence then scanning won't do anything.

- removes the 'reset' option and always resets accounts
- throws an error if the 'from' sequence is not in the chain or the header at
  that sequence is not on the main chain
- sets account heads equal to the block header before the from sequence if it is
  not the genesis block