sets chainProcessor sequence on wallet load (#3660)
* sets chainProcessor sequence on wallet load

the chainProcessor tracks the head of the chain using both the hash and sequence
of a block. when the wallet loads we currently only set the has of the
chainProcessor equal to the latest headd hash from among the accounts in the
wallet.

one consequence of not setting the sequence at load time is that if a user
creates an account while the node is offline, the chainProcessor is not fully
initialized and the newly created account has its head set to null instead of
set equal to the latest head in the wallet. the new account then needs to scan
the whole chain  to sync, even though it was just created and won't have any
transactions earlier in the chain.

- defines getLatestHead to load the latest head (hash and sequence) from among
  all accounts
- sets both the hash and sequence of the chainProcessor from the latest head

* fixes error message for no latest head found