converts account createdAt to store a HeadValue (#3655)
* converts account createdAt to store a HeadValue

we added the createdAt field to AccountValue with a plan in mind to use it for
timestamp-based account birthdays. the createdAt field could be used to speed up
scanning by skipping blocks with a timestamp earlier than the account's
createdAt.

however, we can't rely on block timestamps in this way: we cannot guarantee that
the timestamp for a given block is not far before the wall clock time that that
block was mined at.

to support block-based account birthdays we will store the chain head (block
hash and sequence) at the time the account was created.

- changes the type of createdAt from Date to HeadValue
- updates account import and export for changes to createdAt field

* regenerates fixtures after changing AccountValue