sets createdAt on first on-chain transaction
even if we have an account without a birthday we can use the createdAt field to
save time in future wallet rescans by setting createdAt the first time the
wallet connects a block that contains a transaction for that account.

updates connectBlock to set account.createdAt if the account had a transaction
on the block and createdAt is null. if assetBalanceDeltas is not empty then the
account had a transaction on the block.