enforces enums in wallet transaction rpc schemas
the getAccountTransaction and getAccountTransactions RPC endpoints return a
transaction status and a transaction type with each transaction in a response.
each of these fields has a set of valid values that we define with an enum.

the response schemas, however, only specify that 'status' and 'type' are
strings.

updates the response schemas to enforce that transaction status is one of the
TransactionStatus enum values and that transaction type is one of the
TransactionType enum values.