removes account hasSpend method
when adding a transaction to the wallet we use 'hasSpend' to iterate over all
spends in the transaction to see whether the account spent any notes in the
transaction before using 'addPendingTransaction' or 'connectTransaction'.

however, both of those methods _also_ iterate over all spends in the
transaction. this can result in unnecessary, duplicative iteration over spends.

there is another drawback to the way we currently use 'hasSpend': the account
'addPendingTransaction' and 'connectTransaction' methods depend on the caller
not to call them with any account and transaction where the account was not
involved in the transaction.

- removes 'hasSpend' method
- does not check 'hasSpend' at wallet level before adding transaction to account
- updates account 'addPendingTransaction' and 'connectTransaction' methods to
  return early if the account did not receive or spend any notes in the
  transaction