updates account createdAt if block disconnected (#3675)
* updates account createdAt if block disconnected

when using block-based account birthdays we need to ensure that the account
birthday always refers to a block in the main chain.

on a block disconnect checks whether the account's createdAt field refers to the
disconnected block. if it does, updates the account's createdAt field to the
previous block. this ensures that createdAt refers to a block in the main chain
at the end of a re-org.

- adds conditional update at the end of disconnectBlock

* adds test of update to createdAt with re-org