uses default node account in pool by default (#3697)
by default the name of the account that a mining pool uses for payout
transactions is hardcoded to 'default'. this poses two potential problems:

1. if the default account on the pool node is not 'default', then 'default' may
not have any funds for payouts
2. 'default' may not exist on the pool node

instead of hardcoding the default value of the pool account name we can leave it
unset by default. at the time of a payout the pool can then use the default
account of its node.