Use schema result when routing requests
The bug here is that the schema result is not used when routing
requests. It means that any calls to default() will be ignored in the
resulting request. No transformations will be made.

This fixes it so the transformed input is what is passed to the request
and not the unsanitized output.