IFL-27: resets account birthdays if not in chain (#3679)
* resets account birthdays if latest head not in chain

we need to ensure that the createdAt field of each account always refers to a
block on the main chain (or is null). if we start the node and an account's
createdAt block is not on the node's chain then we can't be sure that it ever
will be (it could be on a fork at any time before or after the node's current
head). in this case we need to reset the account's createdAt to null.

checks that createdAt is on chain for each account on wallet start and resets
account data (including createdAt) if not

* does not reset account birthdays that are ahead of the chain processor