resets account and createdAt during scanning
if, while scanning blocks, we find that an account has a createdAt that refers
to a block on a different fork or chain, reset the account and its createdAt
field and restart the scan.

we need to abort and restart the scan so that we start scanning from genesis for
the newly reset account. otherwise nothing will be scanned for the account until
the node restarts or the user prompts the wallet to start a rescan.