skips scanning blocks based on account birthday (#3685)
* skips scanning blocks based on account birthday

if an account was created at a block that was later than the current scan block
then it is not possible for the account to have any transactions on that block.
we can speed up the scan for that account by skipping decryption for
transactions on that block.

- defines 'shouldScanBlock' to determine whether an account should scan a block
  based on the account's createdAt field
- moves logic for decrpyting notes, connecting transactions on a block to a
  separate wallet method. this was done for readability to avoid deeply nested
  conditional logic in 'connectBlock'

* resets account and createdAt during scanning

if, while scanning blocks, we find that an account has a createdAt that refers
to a block on a different fork or chain, reset the account and its createdAt
field and restart the scan.

we need to abort and restart the scan so that we start scanning from genesis for
the newly reset account. otherwise nothing will be scanned for the account until
the node restarts or the user prompts the wallet to start a rescan.