optionally organizes wallet:transactions output by note
adds a '--notes' flag to the 'wallet:transactions' command. when '--notes' is
passed, each transaction is listed with one row per decryptable output note in
the transaction.

each note row includes the amount of the note, the asset id and name of the
note, and the public addresses of the sender and recipient.

notes are ordered so that notes in the native asset are listed first for each
transaction.

Closes IFL-318 and IFL-409