Wallet block birthdays (#3719)
* converts account createdAt to store a HeadValue (#3655)

* converts account createdAt to store a HeadValue

we added the createdAt field to AccountValue with a plan in mind to use it for
timestamp-based account birthdays. the createdAt field could be used to speed up
scanning by skipping blocks with a timestamp earlier than the account's
createdAt.

however, we can't rely on block timestamps in this way: we cannot guarantee that
the timestamp for a given block is not far before the wall clock time that that
block was mined at.

to support block-based account birthdays we will store the chain head (block
hash and sequence) at the time the account was created.

- changes the type of createdAt from Date to HeadValue
- updates account import and export for changes to createdAt field

* regenerates fixtures after changing AccountValue

* sets imported createdAt to null if not in chain (#3658)

when importing an account with a non-null createdAt block we need to ensure that
that block is on the same chain as the other accounts in the wallet. if it is on
a different chain, on a fork, or ahead of the wallet's chain then we cannot
correctly scan any transactions for that account.

- on import checks whether createdAt refers to a block in the node's chain and
  sets createdAt to null in the new account if not

* updates account createdAt if block disconnected (#3675)

* updates account createdAt if block disconnected

when using block-based account birthdays we need to ensure that the account
birthday always refers to a block in the main chain.

on a block disconnect checks whether the account's createdAt field refers to the
disconnected block. if it does, updates the account's createdAt field to the
previous block. this ensures that createdAt refers to a block in the main chain
at the end of a re-org.

- adds conditional update at the end of disconnectBlock

* adds test of update to createdAt with re-org

* sets createdAt on first on-chain transaction (#3677)

even if we have an account without a birthday we can use the createdAt field to
save time in future wallet rescans by setting createdAt the first time the
wallet connects a block that contains a transaction for that account.

updates connectBlock to set account.createdAt if the account had a transaction
on the block and createdAt is null. if assetBalanceDeltas is not empty then the
account had a transaction on the block.

* migrates account createdAt from Date to HeadValue (#3656)

* migrates account createdAt from Date to HeadValue

to support block-based account birthdays we will store the chain head (block
hash and sequence) at the time the account was created.

adds migration to set createdAt to null for all existing accounts

* implements backward migration

* adds comments for numeric literals in encodings

* IFL-27: resets account birthdays if not in chain (#3679)

* resets account birthdays if latest head not in chain

we need to ensure that the createdAt field of each account always refers to a
block on the main chain (or is null). if we start the node and an account's
createdAt block is not on the node's chain then we can't be sure that it ever
will be (it could be on a fork at any time before or after the node's current
head). in this case we need to reset the account's createdAt to null.

checks that createdAt is on chain for each account on wallet start and resets
account data (including createdAt) if not

* does not reset account birthdays that are ahead of the chain processor

* skips scanning blocks based on account birthday (#3685)

* skips scanning blocks based on account birthday

if an account was created at a block that was later than the current scan block
then it is not possible for the account to have any transactions on that block.
we can speed up the scan for that account by skipping decryption for
transactions on that block.

- defines 'shouldScanBlock' to determine whether an account should scan a block
  based on the account's createdAt field
- moves logic for decrpyting notes, connecting transactions on a block to a
  separate wallet method. this was done for readability to avoid deeply nested
  conditional logic in 'connectBlock'

* resets account and createdAt during scanning

if, while scanning blocks, we find that an account has a createdAt that refers
to a block on a different fork or chain, reset the account and its createdAt
field and restart the scan.

we need to abort and restart the scan so that we start scanning from genesis for
the newly reset account. otherwise nothing will be scanned for the account until
the node restarts or the user prompts the wallet to start a rescan.

* fixes account fixtures for genesis.test.slow

* regenerates accouint fixtures for new tests