adds transactionHashToBlockHash chain index
adds an index mapping transaction hash to block hash for all transactions on the
main chain. the index will support looking up transactions on the chain using
only the transaction hash. this lookup currently requires a block hash because
the transactions store is keyed by block hash.

the index is a datastore with a TransactionHash (i.e., Buffer) key and a
BlockHash (i.e., Buffer) value.

entries are inserted into the index for each transaction when a block is
connected to the main chain (in saveConnect). entries are deleted from the index
for each transaction when a block is disconnected from the main chain (in
saveDisconnect).