adds optional spends to wallet/createTransaction
supports custom funding for transactions by taking a list of 'spends' where each
spend is a serialized note and the index of the note.

if spends are passed in a request to wallet/createTransaction the endpoint uses
the chain to create a witness for the note at that index. the note and witness
are then passed to the wallet createTransaction method. if spends are passed to
that method then it bypasses funding the transaction by looking up notes in the
wallet database.