adds markSpent parameter to createTransaction RPC
optionally marks notes used in creating a transaction as spent. this is useful
for creating multiple raw transactions in succession without spending the same
note(s) in more than one transaction.

we have an example of this use case in the airdrop: we need to create many
raw transactions and then post them separately before broadcasting to the
network.