adds unspentNotesCount to getBalance(s) response
it may be useful for some developers to know how many unspent notes an account
has.

we already iterate over unspent notes and sum their values when calculating an
account's available balance. by also counting the number of notes during this
calculation we can add the unspent notes count to the responses of
wallet/getBalance and wallet/getBalances