allows docker deploy with version tags only (#3792)
we recently added the ability to deploy docker images tagged with the version
tag (e.g., v0.1.76)

however, the workflow currently only deploys these tags if we are also deploying
to and overwriting the 'latest' tag. this prevents us from deploying a tag from
an earlier version, like v0.1.75, without overwriting the 'latest' image from
the current version

we also use the 'github.event' context to determine if we're deploying a tag.
this is incorrect since the only event that can trigger this workflow is
'workflow_dispatch', which does not have a 'ref_type' parameter, so
`github.event.ref_type == 'tag'` will never be true.  instead we can use
'github.ref_type' which will refer to the branch or tag that the workflow was
triggered from.