displays transactions sent to own address in wallet:transactions (#3799)
we exclude IRON from the transactions list in a given transaction if the delta
was zero. however, if an account sends a transaction in IRON to its own address
then this transaction is omitted from the list completely.

to fix this we can ensure that we only exclude a row for IRON if it is not the
only asset in the transaction.