clears walletdb datastores during reset (#3797)
* clears walletdb datastores during reset

following a network reset the walletdb may have data that is incompatible with
database schemas. for instance, if asset identifiers in transactions or notes
change, then any datastores that encode that data must be migrated or cleared.

we avoid deleting the walletdb during 'reset' so that we do not delete user
keys. however, the node will fail to start if it needs to apply migrations on
incompatible data or even if it needs to deserialize incompatible data in the
course of deleting it.

to resolve this we can use 'clear' on each walletdb datastore during the 'reset'
command. clear will delete all keys within a range without deserializing the
values.

note that clear soft-deletes data that is cleanup up later during database
compactions rather than deleting immediately.

- clears each walletdb datastore, except accounts and meta, during 'reset'
- adds missing datastores to cleanupDeletedAccounts

* defines list of walletDb 'cacheStores'

this is the list of stores that cache and/or index decrypted chain data that
must be cleared when resetting the wallet or deleting an account