adds support for loading unspent notes in different orderings
by default unspent notes for a given account and asset are always loaded in
order of increasing block sequence.

it may be advantageous sometimes to load notes by value so that transactions can
be funded with larger notes first.

renames 'loadUnspentNoteHashes' and 'loadUnspentNotes' to reflect that they are
loaded by sequence.

adds walletDb methods to support loading unspent notes in order of value instead
of sequence (i.e., 'loadUnspentNoteHashesByValue').

adds 'reverse' parameter to each of these methods to support loading in reverse
order.