adds custom transaction funding by note hash
supporting custom transaction funding will allow users and developers to choose
the notes to spend in a transaction using whatever note selection algorithm they
choose.

adds optional 'spendNoteHashes' parameter to 'wallet/createTransaction' rpc
endpoint to accept a list of hex string note hashes to use to fund a
transaction.

implements 'fundWithNoteHashes' to load decrypted notes from the specified
hashes and use them to build the spends for a raw transaction. ensures that the
specified notes have enough value to fund the transaction. requires that notes
have an index so that a witness can be created.

does not check whether the notes have already been spent. this is intentional to
support use cases like 'canceling' or 'speeding up' a transaction by creating a
transaction that spends the same notes but offers a higher fee.

adds unit tests for wallet and rpc transaction creation methods.