supports partial funding from notes passed to createTransaction
if the notes passed to createTransaction do not cover the amount needed, the
wallet will fund the rest of the transaction using available unspent notes.

refactors fund to optionally take a list of notes.

renames 'createSpendsForAsset' to 'addSpendsForAsset' and modifies to take a set
of notes already spent and an amount already spent.

removes unneeded code: 'checkNoteSpentOnChainAndRepair', 'createSpends'

renames 'spendNoteHashes' to 'notes' in RPC and createTransaction method